chore: reorganized render code
