+++ include/bitmap.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 
 void write_bitmap(char *filename, unsigned int imgwidth, unsigned int imgheight,
-                  uint8_t pixels[][3]);
+                  uint8_t *pixels);
